remove oltp
