+++ crates/log/src/lib.rs
@@ -1,9 +1,4 @@
-use opentelemetry::trace::TracerProvider;
 use opentelemetry_gcloud_trace::errors::GcloudTraceError;
-use opentelemetry_gcloud_trace::{GcpCloudTraceExporterBuilder, SdkTracer};
-use opentelemetry_otlp::SpanExporterBuilder;
-use opentelemetry_sdk::trace::{RandomIdGenerator, SdkTracerProvider};
-use opentelemetry_sdk::Resource;
 use tracing::subscriber::SetGlobalDefaultError;
 use tracing_log::log::SetLoggerError;
 use tracing_subscriber::layer::SubscriberExt;
@@ -17,19 +12,8 @@ pub mod otel;
 pub use fmt::LogFormat;
 
 #[derive(Debug, Clone)]
-pub enum Tracer {
-    Otlp(OtlpConfig),
-    Gcloud(GcloudConfig),
-}
-
-#[derive(Debug, Clone)]
-pub struct OtlpConfig {
-    pub endpoint: Option<String>,
-}
-
-#[derive(Debug, Clone)]
-pub struct GcloudConfig {
-    pub project_id: Option<String>,
+pub enum TracerConfig {
+    Gcloud(gcloud::GcloudConfig),
 }
 
 #[derive(Debug, thiserror::Error)]
@@ -45,12 +29,15 @@ pub enum Error {
 
     #[error("google cloud trace error: {0}")]
     GcloudTrace(#[from] GcloudTraceError),
+
+    #[error("failed to install crypto provider")]
+    InstallCryptoFailed,
 }
 
 pub async fn init(
     format: LogFormat,
     dev_log: bool,
-    telemetry_config: Option<Tracer>,
+    telemetry_config: Option<TracerConfig>,
 ) -> Result<(), Error> {
     const DEFAULT_LOG_FILTER: &str = "cairo_native::compiler=off,pipeline=debug,stage=debug,info,\
                                       tasks=debug,executor=trace,forking::backend=trace,\
@@ -63,10 +50,6 @@ pub async fn init(
         DEFAULT_LOG_FILTER.to_string()
     };
 
-    rustls::crypto::ring::default_provider()
-        .install_default()
-        .expect("Failed to install rustls crypto provider");
-
     // If the user has set the `RUST_LOG` environment variable, then we prioritize it.
     // Otherwise, we use the default log filter.
     // TODO: change env var to `KATANA_LOG`.
@@ -76,12 +59,8 @@ pub async fn init(
     if let Some(telemetry_config) = telemetry_config {
         // Initialize telemetry layer based on exporter type
         let telemetry = match telemetry_config {
-            Tracer::Gcloud(cfg) => {
-                let tracer = init_gcp_tracer(&cfg).await?;
-                tracing_opentelemetry::layer().with_tracer(tracer)
-            }
-            Tracer::Otlp(cfg) => {
-                let tracer = init_otlp_tracer(&cfg)?;
+            TracerConfig::Gcloud(cfg) => {
+                let tracer = gcloud::init_gcp_tracer(&cfg).await?;
                 tracing_opentelemetry::layer().with_tracer(tracer)
             }
         };
@@ -102,56 +81,4 @@ pub async fn init(
     }
 
     Ok(())
-}
-
-/// Initialize Google Cloud Trace exporter
-///
-/// Make sure to set `GOOGLE_APPLICATION_CREDENTIALS` env var to authenticate to gcloud
-async fn init_gcp_tracer(gcloud_config: &GcloudConfig) -> Result<SdkTracer, Error> {
-    let resource = Resource::builder().with_service_name("katana").build();
-
-    let mut trace_exporter = if let Some(project_id) = &gcloud_config.project_id {
-        GcpCloudTraceExporterBuilder::new(project_id.clone())
-    } else {
-        // Default will attempt to find project ID from environment variables in the following
-        // order:
-        // - GCP_PROJECT
-        // - PROJECT_ID
-        // - GCP_PROJECT_ID
-        GcpCloudTraceExporterBuilder::for_default_project_id().await?
-    };
-
-    trace_exporter = trace_exporter.with_resource(resource);
-
-    let tracer_provider = trace_exporter.create_provider().await?;
-    let tracer = trace_exporter.install(&tracer_provider).await.unwrap();
-
-    opentelemetry::global::set_tracer_provider(tracer_provider.clone());
-
-    Ok(tracer)
-}
-
-/// Initialize OTLP tracer
-fn init_otlp_tracer(otlp_config: &OtlpConfig) -> Result<opentelemetry_sdk::trace::Tracer, Error> {
-    use opentelemetry_otlp::WithExportConfig;
-
-    let resource = Resource::builder().with_service_name("katana").build();
-
-    let mut exporter_builder = SpanExporterBuilder::new().with_tonic();
-
-    if let Some(endpoint) = &otlp_config.endpoint {
-        exporter_builder = exporter_builder.with_endpoint(endpoint);
-    }
-
-    let exporter = exporter_builder.build().unwrap();
-
-    let provider = SdkTracerProvider::builder()
-        .with_id_generator(RandomIdGenerator::default())
-        .with_batch_exporter(exporter)
-        .with_resource(resource)
-        .build();
-
-    opentelemetry::global::set_tracer_provider(provider.clone());
-
-    Ok(provider.tracer("katana"))
 }
