+++ crates/rpc/rpc/Cargo.toml
@@ -25,7 +25,6 @@ anyhow.workspace = true
 cainome = { workspace = true, optional = true }
 futures.workspace = true
 http.workspace = true
-http-body.workspace = true
 jsonrpsee = { workspace = true, features = [ "server" ] }
 metrics.workspace = true
 reqwest = { workspace = true, optional = true }
